Add phoneId to SMS/MMS tables
This is an interim change to add phoneId for Phase 1 of MMS
app enablement. This will be later removed as AOSP tables
support subId

Initial reason for adding phoneId in L code:
As per AOSP code, SMS/MMS are storing with SUB ID in database
due to this if user changed the SIM's then user won't be able
to see the already stored SMS/MMS in phone.
Modifying data base to store SMS/MMS with PHONE ID, now user
will be able to see the stored MMS/SMS in phone even after
inserting the different SIM's

Change-Id: I066381564e1fbfc5002b453d832801eaeeec0836